Refactor file_upload into package
